+++ benchmarks/index.js
@@ -1,8 +1,8 @@
 import { h } from 'preact';
 import Suite from 'benchmarkjs-pretty';
-import renderToStringBaseline from './lib/render-to-string';
+import renderToStringBaseline from 'baseline-rts';
 // import renderToString from '../src/index';
-import renderToString from '../dist/index.mjs';
+import renderToString from '../dist/index.module.js';
 import TextApp from './text';
 // import StackApp from './stack';
 import { App as IsomorphicSearchResults } from './isomorphic-ui/search-results/index';