+++ src/index.js
@@ -1,16 +1,12 @@
 import {
 	encodeEntities,
 	styleObjToCss,
-	getContext,
-	createComponent,
 	transformAttributeName,
-	UNSAFE_NAME,
-	XLINK,
-	VOID_ELEMENTS
+	UNSAFE_NAME
 } from './util';
 import { options, h, Fragment } from 'preact';
-import { _renderToStringPretty } from './pretty';
 import {
+	CHILDREN,
 	COMMIT,
 	COMPONENT,
 	DIFF,
@@ -20,40 +16,25 @@ import {
 	PARENT,
 	RENDER,
 	SKIP_EFFECTS,
-	VNODE,
-	CHILDREN
+	VNODE
 } from './constants';
 
 /** @typedef {import('preact').VNode} VNode */
 
-const SHALLOW = { shallow: true };
-
-/** Render Preact JSX + Components to an HTML string.
- *	@name render
- *	@function
- *	@param {VNode} vnode	JSX VNode to render.
- *	@param {Object} [context={}]	Optionally pass an initial context object through the render path.
- *	@param {Object} [options={}]	Rendering options
- *	@param {Boolean} [options.shallow=false]	If `true`, renders nested Components as HTML elements (`<Foo a="b" />`).
- *	@param {Boolean} [options.xml=false]		If `true`, uses self-closing tags for elements without children.
- *	@param {Boolean} [options.pretty=false]		If `true`, adds whitespace for readability
- *	@param {RegExp|undefined} [options.voidElements]       RegeEx that matches elements that are considered void (self-closing)
- */
-renderToString.render = renderToString;
-
-/** Only render elements, leaving Components inline as `<ComponentName ... />`.
- *	This method is just a convenience alias for `render(vnode, context, { shallow:true })`
- *	@name shallow
- *	@function
- *	@param {VNode} vnode	JSX VNode to render.
- *	@param {Object} [context={}]	Optionally pass an initial context object through the render path.
- */
-let shallowRender = (vnode, context) => renderToString(vnode, context, SHALLOW);
-
 const EMPTY_ARR = [];
-function renderToString(vnode, context, opts) {
-	context = context || {};
+const isArray = Array.isArray;
+const assign = Object.assign;
 
+// Global state for the current render pass
+let beforeDiff, afterDiff, renderHook, ummountHook;
+
+/**
+ * Render Preact JSX + Components to an HTML string.
+ * @param {VNode} vnode	JSX Element / VNode to render
+ * @param {Object} [context={}] Initial root context object
+ * @returns {string} serialized HTML
+ */
+export default function renderToString(vnode, context) {
 	// Performance optimization: `renderToString` is synchronous and we
 	// therefore don't execute any effects. To do that we pass an empty
 	// array to `options._commit` (`__c`). But we can go one step further
@@ -62,83 +43,67 @@ function renderToString(vnode, context, opts) {
 	const previousSkipEffects = options[SKIP_EFFECTS];
 	options[SKIP_EFFECTS] = true;
 
+	// store options hooks once before each synchronous render call
+	beforeDiff = options[DIFF];
+	afterDiff = options[DIFFED];
+	renderHook = options[RENDER];
+	ummountHook = options.unmount;
+
 	const parent = h(Fragment, null);
 	parent[CHILDREN] = [vnode];
 
-	let res;
-	if (
-		opts &&
-		(opts.pretty ||
-			opts.voidElements ||
-			opts.sortAttributes ||
-			opts.shallow ||
-			opts.allAttributes ||
-			opts.xml ||
-			opts.attributeHook)
-	) {
-		res = _renderToStringPretty(vnode, context, opts);
-	} else {
-		res = _renderToString(vnode, context, false, undefined, parent);
+	try {
+		return _renderToString(
+			vnode,
+			context || EMPTY_OBJ,
+			false,
+			undefined,
+			parent
+		);
+	} finally {
+		// options._commit, we don't schedule any effects in this library right now,
+		// so we can pass an empty queue to this hook.
+		if (options[COMMIT]) options[COMMIT](vnode, EMPTY_ARR);
+		options[SKIP_EFFECTS] = previousSkipEffects;
+		EMPTY_ARR.length = 0;
 	}
-
-	// options._commit, we don't schedule any effects in this library right now,
-	// so we can pass an empty queue to this hook.
-	if (options[COMMIT]) options[COMMIT](vnode, EMPTY_ARR);
-	options[SKIP_EFFECTS] = previousSkipEffects;
-	EMPTY_ARR.length = 0;
-
-	return res;
 }
 
-function renderFunctionComponent(vnode, context) {
-	let rendered,
-		c = createComponent(vnode, context),
-		cctx = getContext(vnode.type, context);
-
-	vnode[COMPONENT] = c;
-
-	// If a hook invokes setState() to invalidate the component during rendering,
-	// re-render it up to 25 times to allow "settling" of memoized states.
-	// Note:
-	//   This will need to be updated for Preact 11 to use internal.flags rather than component._dirty:
-	//   https://github.com/preactjs/preact/blob/d4ca6fdb19bc715e49fd144e69f7296b2f4daa40/src/diff/component.js#L35-L44
-	let renderHook = options[RENDER];
-	let count = 0;
-	while (c[DIRTY] && count++ < 25) {
-		c[DIRTY] = false;
-
-		if (renderHook) renderHook(vnode);
-
-		// stateless functional components
-		rendered = vnode.type.call(c, vnode.props, cctx);
-	}
-
-	return rendered;
+// Installed as setState/forceUpdate for function components
+function markAsDirty() {
+	this.__d = true;
 }
 
+const EMPTY_OBJ = {};
+
+/**
+ * @param {VNode} vnode
+ * @param {Record<string, unknown>} context
+ */
 function renderClassComponent(vnode, context) {
-	let nodeName = vnode.type,
-		cctx = getContext(nodeName, context);
+	let type = /** @type {import("preact").ComponentClass<typeof vnode.props>} */ (vnode.type);
 
-	// c = new nodeName(props, context);
-	let c = new nodeName(vnode.props, cctx);
+	let c = new type(vnode.props, context);
+
 	vnode[COMPONENT] = c;
 	c[VNODE] = vnode;
+
+	c.props = vnode.props;
+	c.context = context;
 	// turn off stateful re-rendering:
 	c[DIRTY] = true;
-	c.props = vnode.props;
-	if (c.state == null) c.state = {};
+
+	if (c.state == null) c.state = EMPTY_OBJ;
 
 	if (c[NEXT_STATE] == null) {
 		c[NEXT_STATE] = c.state;
 	}
 
-	c.context = cctx;
-	if (nodeName.getDerivedStateFromProps) {
+	if (type.getDerivedStateFromProps) {
 		c.state = assign(
 			{},
 			c.state,
-			nodeName.getDerivedStateFromProps(c.props, c.state)
+			type.getDerivedStateFromProps(c.props, c.state)
 		);
 	} else if (c.componentWillMount) {
 		c.componentWillMount();
@@ -148,46 +113,20 @@ function renderClassComponent(vnode, context) {
 		c.state = c[NEXT_STATE] !== c.state ? c[NEXT_STATE] : c.state;
 	}
 
-	let renderHook = options[RENDER];
 	if (renderHook) renderHook(vnode);
 
-	return c.render(c.props, c.state, c.context);
+	return c.render(c.props, c.state, context);
 }
 
-function normalizePropName(name, isSvgMode) {
-	if (name === 'className') {
-		return 'class';
-	} else if (name === 'htmlFor') {
-		return 'for';
-	} else if (name === 'defaultValue') {
-		return 'value';
-	} else if (name === 'defaultChecked') {
-		return 'checked';
-	} else if (name === 'defaultSelected') {
-		return 'selected';
-	} else if (isSvgMode && XLINK.test(name)) {
-		return name.toLowerCase().replace(/^xlink:?/, 'xlink:');
-	}
-
-	return name;
-}
-
-function normalizePropValue(name, v) {
-	if (name === 'style' && v != null && typeof v === 'object') {
-		return styleObjToCss(v);
-	} else if (name[0] === 'a' && name[1] === 'r' && typeof v === 'boolean') {
-		// always use string values instead of booleans for aria attributes
-		// also see https://github.com/preactjs/preact/pull/2347/files
-		return String(v);
-	}
-
-	return v;
-}
-
-const isArray = Array.isArray;
-const assign = Object.assign;
-
-/** The default export is an alias of `render()`. */
+/**
+ * Recursively render VNodes to HTML.
+ * @param {VNode|any} vnode
+ * @param {any} context
+ * @param {boolean} isSvgMode
+ * @param {any} selectValue
+ * @param {VNode} parent
+ * @returns {string}
+ */
 function _renderToString(vnode, context, isSvgMode, selectValue, parent) {
 	// Ignore non-rendered VNodes/values
 	if (vnode == null || vnode === true || vnode === false || vnode === '') {
@@ -197,7 +136,7 @@ function _renderToString(vnode, context, isSvgMode, selectValue, parent) {
 	// Text VNodes: escape as HTML
 	if (typeof vnode !== 'object') {
 		if (typeof vnode === 'function') return '';
-		return encodeEntities(vnode);
+		return encodeEntities(vnode + '');
 	}
 
 	// Recurse into children / Arrays
@@ -205,9 +144,12 @@ function _renderToString(vnode, context, isSvgMode, selectValue, parent) {
 		let rendered = '';
 		parent[CHILDREN] = vnode;
 		for (let i = 0; i < vnode.length; i++) {
+			let child = vnode[i];
+			if (child == null || typeof child === 'boolean') continue;
+
 			rendered =
 				rendered +
-				_renderToString(vnode[i], context, isSvgMode, selectValue, parent);
+				_renderToString(child, context, isSvgMode, selectValue, parent);
 		}
 		return rendered;
 	}
@@ -216,26 +158,60 @@ function _renderToString(vnode, context, isSvgMode, selectValue, parent) {
 	if (vnode.constructor !== undefined) return '';
 
 	vnode[PARENT] = parent;
-	if (options[DIFF]) options[DIFF](vnode);
+	if (beforeDiff) beforeDiff(vnode);
 
 	let type = vnode.type,
-		props = vnode.props;
+		props = vnode.props,
+		cctx = context,
+		contextType,
+		rendered,
+		component;
 
 	// Invoke rendering on Components
-	const isComponent = typeof type === 'function';
-	if (isComponent) {
-		let rendered;
+	if (typeof type === 'function') {
 		if (type === Fragment) {
 			rendered = props.children;
 		} else {
-			if (type.prototype && typeof type.prototype.render === 'function') {
-				rendered = renderClassComponent(vnode, context);
-			} else {
-				rendered = renderFunctionComponent(vnode, context);
+			contextType = type.contextType;
+			if (contextType != null) {
+				let provider = context[contextType.__c];
+				cctx = provider ? provider.props.value : contextType.__;
 			}
 
-			let component = vnode[COMPONENT];
-			if (component.getChildContext) {
+			if (type.prototype && typeof type.prototype.render === 'function') {
+				rendered = /**#__NOINLINE__**/ renderClassComponent(vnode, cctx);
+				component = vnode[COMPONENT];
+			} else {
+				component = {
+					__v: vnode,
+					props,
+					context: cctx,
+					// silently drop state updates
+					setState: markAsDirty,
+					forceUpdate: markAsDirty,
+					__d: true,
+					// hooks
+					__h: []
+				};
+				vnode[COMPONENT] = component;
+
+				// If a hook invokes setState() to invalidate the component during rendering,
+				// re-render it up to 25 times to allow "settling" of memoized states.
+				// Note:
+				//   This will need to be updated for Preact 11 to use internal.flags rather than component._dirty:
+				//   https://github.com/preactjs/preact/blob/d4ca6fdb19bc715e49fd144e69f7296b2f4daa40/src/diff/component.js#L35-L44
+				let count = 0;
+				while (component[DIRTY] && count++ < 25) {
+					component[DIRTY] = false;
+
+					if (renderHook) renderHook(vnode);
+
+					rendered = type.call(component, props, cctx);
+				}
+				component[DIRTY] = true;
+			}
+
+			if (component.getChildContext != null) {
 				context = assign({}, context, component.getChildContext());
 			}
 		}
@@ -254,156 +230,156 @@ function _renderToString(vnode, context, isSvgMode, selectValue, parent) {
 			selectValue,
 			vnode
 		);
-
-		if (options[DIFFED]) options[DIFFED](vnode);
+		if (afterDiff) afterDiff(vnode);
 		vnode[PARENT] = undefined;
 
-		if (options.unmount) options.unmount(vnode);
+		if (ummountHook) ummountHook(vnode);
 
 		return str;
 	}
 
 	// Serialize Element VNodes to HTML
-	let s = '<',
-		children,
-		html;
+	let s = '<' + type,
+		html = '',
+		children;
 
-	s = s + type;
+	for (let name in props) {
+		let v = props[name];
 
-	if (props) {
-		children = props.children;
-		for (let name in props) {
-			let v = props[name];
-
-			if (
-				name === 'key' ||
-				name === 'ref' ||
-				name === '__self' ||
-				name === '__source' ||
-				name === 'children' ||
-				(name === 'className' && 'class' in props) ||
-				(name === 'htmlFor' && 'for' in props)
-			) {
-				continue;
-			}
-
-			if (UNSAFE_NAME.test(name)) continue;
-
-			name = normalizePropName(name, isSvgMode);
-			v = normalizePropValue(name, v);
-
-			if (name === 'dangerouslySetInnerHTML') {
-				html = v && v.__html;
-			} else if (type === 'textarea' && name === 'value') {
-				// <textarea value="a&b"> --> <textarea>a&amp;b</textarea>
+		switch (name) {
+			case 'children':
 				children = v;
-			} else if ((v || v === 0 || v === '') && typeof v !== 'function') {
-				name = transformAttributeName(name);
+				continue;
 
-				if (v === true || v === '') {
-					v = name;
-					s = s + ' ' + name;
-					continue;
-				}
+			// VDOM-specific props
+			case 'key':
+			case 'ref':
+			case '__self':
+			case '__source':
+				continue;
 
-				if (name === 'value') {
-					if (type === 'select') {
+			// prefer for/class over htmlFor/className
+			case 'htmlFor':
+				if ('for' in props) continue;
+				name = 'for';
+				break;
+			case 'className':
+				if ('class' in props) continue;
+				name = 'class';
+				break;
+
+			// Form element reflected properties
+			case 'defaultChecked':
+				name = 'checked';
+				break;
+			case 'defaultSelected':
+				name = 'selected';
+				break;
+
+			// Special value attribute handling
+			case 'defaultValue':
+			case 'value':
+				name = 'value';
+				switch (type) {
+					// <textarea value="a&b"> --> <textarea>a&amp;b</textarea>
+					case 'textarea':
+						children = v;
+						continue;
+
+					// <select value> is serialized as a selected attribute on the matching option child
+					case 'select':
 						selectValue = v;
 						continue;
-					} else if (
-						// If we're looking at an <option> and it's the currently selected one
-						type === 'option' &&
-						selectValue == v &&
-						// and the <option> doesn't already have a selected attribute on it
-						!('selected' in props)
-					) {
-						s = s + ' selected';
-					}
+
+					// Add a selected attribute to <option> if its value matches the parent <select> value
+					case 'option':
+						if (selectValue == v && !('selected' in props)) {
+							s = s + ' selected';
+						}
+						break;
 				}
-				s = s + ' ' + name + '="' + encodeEntities(v) + '"';
+				break;
+
+			case 'dangerouslySetInnerHTML':
+				html = v && v.__html;
+				continue;
+
+			// serialize object styles to a CSS string
+			case 'style':
+				if (typeof v === 'object') {
+					v = styleObjToCss(v);
+				}
+				break;
+
+			default: {
+				if (UNSAFE_NAME.test(name)) {
+					continue;
+				} else if ((name[4] === '-' || name === 'draggable') && v != null) {
+					// serialize boolean aria-xyz or draggable attribute values as strings
+					// `draggable` is an enumerated attribute and not Boolean. A value of `true` or `false` is mandatory
+					// https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/draggable
+					v += '';
+				} else {
+					name = transformAttributeName(name);
+				}
+			}
+		}
+
+		// write this attribute to the buffer
+		if (v != null && v !== false && typeof v !== 'function') {
+			if (v === true || v === '') {
+				s = s + ' ' + name;
+			} else {
+				s = s + ' ' + name + '="' + encodeEntities(v + '') + '"';
 			}
 		}
 	}
-
-	let startElement = s;
-	s = s + '>';
 
 	if (UNSAFE_NAME.test(type)) {
-		throw new Error(`${type} is not a valid HTML tag name in ${s}`);
+		// this seems to performs a lot better than throwing
+		// return '<!-- -->';
+		throw new Error(`${type} is not a valid HTML tag name in ${s}>`);
 	}
-
-	let pieces = '';
-	let hasChildren = false;
 
 	if (html) {
-		pieces = pieces + html;
-		hasChildren = true;
+		// dangerouslySetInnerHTML defined this node's contents
 	} else if (typeof children === 'string') {
-		pieces = pieces + encodeEntities(children);
-		hasChildren = true;
-	} else if (isArray(children)) {
-		vnode[CHILDREN] = children;
-		for (let i = 0; i < children.length; i++) {
-			let child = children[i];
-			if (child != null && child !== false) {
-				let childSvgMode =
-					type === 'svg' || (type !== 'foreignObject' && isSvgMode);
-				let ret = _renderToString(
-					child,
-					context,
-					childSvgMode,
-					selectValue,
-					vnode
-				);
-
-				// Skip if we received an empty string
-				if (ret) {
-					pieces = pieces + ret;
-					hasChildren = true;
-				}
-			}
-		}
+		// single text child
+		html = encodeEntities(children);
 	} else if (children != null && children !== false && children !== true) {
-		vnode[CHILDREN] = [children];
+		// recurse into this element VNode's children
 		let childSvgMode =
 			type === 'svg' || (type !== 'foreignObject' && isSvgMode);
-		let ret = _renderToString(
-			children,
-			context,
-			childSvgMode,
-			selectValue,
-			vnode
-		);
-
-		// Skip if we received an empty string
-		if (ret) {
-			pieces = pieces + ret;
-			hasChildren = true;
-		}
+		html = _renderToString(children, context, childSvgMode, selectValue, vnode);
 	}
 
-	if (options[DIFFED]) options[DIFFED](vnode);
+	if (afterDiff) afterDiff(vnode);
 	vnode[PARENT] = undefined;
-	if (options.unmount) options.unmount(vnode);
+	if (ummountHook) ummountHook(vnode);
 
-	if (hasChildren) {
-		s = s + pieces;
-	} else if (VOID_ELEMENTS.test(type)) {
-		return startElement + ' />';
+	// Emit self-closing tag for empty void elements:
+	if (!html && SELF_CLOSING.has(type)) {
+		return s + '/>';
 	}
 
-	return s + '</' + type + '>';
+	return s + '>' + html + '</' + type + '>';
 }
 
-/** The default export is an alias of `render()`. */
-
-renderToString.shallowRender = shallowRender;
-
-export default renderToString;
-
-export {
-	renderToString as render,
-	renderToString as renderToStaticMarkup,
-	renderToString,
-	shallowRender
-};
+const SELF_CLOSING = new Set([
+	'area',
+	'base',
+	'br',
+	'col',
+	'command',
+	'embed',
+	'hr',
+	'img',
+	'input',
+	'keygen',
+	'link',
+	'meta',
+	'param',
+	'source',
+	'track',
+	'wbr'
+]);