+++ CHANGELOG.md
@@ -1,5 +1,50 @@
 # preact-render-to-string
 
+## 6.0.1
+
+### Patch Changes
+
+- [#292](https://github.com/preactjs/preact-render-to-string/pull/292) [`8f4692c`](https://github.com/preactjs/preact-render-to-string/commit/8f4692c49277591819acb74808a0e28f7cb30c2f) Thanks [@marvinhagemeister](https://github.com/marvinhagemeister)! - Fix error in commonjs entry point
+
+## 6.0.0
+
+### Major Changes
+
+- [#241](https://github.com/preactjs/preact-render-to-string/pull/241) [`e8cbf66`](https://github.com/preactjs/preact-render-to-string/commit/e8cbf66b2620842671b5c95817454ff2ddb0e450) Thanks [@developit](https://github.com/developit)! - Improve performance by another 5-10% using `switch` and short-circuiting, and move pretty-printing from into `preact-render-to-string/jsx`.
+
+* [#282](https://github.com/preactjs/preact-render-to-string/pull/282) [`6376f62`](https://github.com/preactjs/preact-render-to-string/commit/6376f62309ec19482ded68406a03910ad6de57d1) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Remove trailing space for void_elements, this could fail some test_assertions as
+  `<img />` will become `<img/>`, the other `VOID_ELEMENTS` this will be applied for
+  can be found [here](https://github.com/preactjs/preact-render-to-string/blob/remove-trailing-space/src/index.js#L368-L385)
+
+- [#286](https://github.com/preactjs/preact-render-to-string/pull/286) [`7a8b590`](https://github.com/preactjs/preact-render-to-string/commit/7a8b590237a70e8708e46cb1d92ab24327f60160) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Remove the castin to VNode for `preact/debug`, this is fixed in Preact >= 10.13.0
+
+### Patch Changes
+
+- [#286](https://github.com/preactjs/preact-render-to-string/pull/286) [`7a8b590`](https://github.com/preactjs/preact-render-to-string/commit/7a8b590237a70e8708e46cb1d92ab24327f60160) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Change style calculation to use a Set rather than Regex
+
+* [#285](https://github.com/preactjs/preact-render-to-string/pull/285) [`a0546fe`](https://github.com/preactjs/preact-render-to-string/commit/a0546fe6a008eca54edfcd97f775c4b5581d086c) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Fix CJS export
+
+- [#288](https://github.com/preactjs/preact-render-to-string/pull/288) [`0b04860`](https://github.com/preactjs/preact-render-to-string/commit/0b0486029ca10a846c368a0ab404e4b2e8eb96d6) Thanks [@glenchao](https://github.com/glenchao)! - Enumerate draggable attribute, so the output isn't `draggable` but `draggable="true"`
+
+* [#283](https://github.com/preactjs/preact-render-to-string/pull/283) [`3defa9d`](https://github.com/preactjs/preact-render-to-string/commit/3defa9dab48cc7379c9135d92273fbb7f886cab0) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Follow up fixes to #278
+
+- [#289](https://github.com/preactjs/preact-render-to-string/pull/289) [`07ebc66`](https://github.com/preactjs/preact-render-to-string/commit/07ebc66bb76b775cae58b5163acf2943ec72c8bf) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Support `data` attribute
+
+* [#270](https://github.com/preactjs/preact-render-to-string/pull/270) [`5c6877d`](https://github.com/preactjs/preact-render-to-string/commit/5c6877d13d60b4cdd87632ac3052b006207568ff) Thanks [@developit](https://github.com/developit)! - improve unmount option hook call performance
+
+- [#278](https://github.com/preactjs/preact-render-to-string/pull/278) [`8cf7cef`](https://github.com/preactjs/preact-render-to-string/commit/8cf7cef0e96b3e48ffea5fcf4f76db6410de8346) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Improve performance by
+
+  - storing the void_elements in a Set
+  - hoisting the `x-link` regex
+  - remove case-insensitive from regexes and calling `.toLowerCase()` instead
+  - caching suffixes for css-props
+
+## 5.2.6
+
+### Patch Changes
+
+- [#257](https://github.com/preactjs/preact-render-to-string/pull/257) [`8b944b2`](https://github.com/preactjs/preact-render-to-string/commit/8b944b28be64d470a947f999153c9b64b078f7a8) Thanks [@marvinhagemeister](https://github.com/marvinhagemeister)! - Fix `preact/debug` incorrectly throwing errors on text children
+
 ## 5.2.5
 
 ### Patch Changes